requirements.txt
streamlit
openai
pandas
Pillow
requests